pyproject.toml
[project]
name = "repl-nix-thermaldynamicsvisualizer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.9.2",
    "numpy>=2.1.2",
    "pillow>=10.4.0",
    "plotly>=5.24.1",
    "scipy>=1.14.1",
    "streamlit>=1.39.0",
]
